feat(big): restructure llm interactions
1. use DSPy to organise prompting better
2. split llm completions for trade decision, data check and data request
3. enable having multiple traders, and finding best decisions from there (to be iterated/improved on)
4. convert intput and outputs from llm trder class to classes
5. del llm_interface
6. rm litellm (no longer needed, DSPy itself supports many models)

+++ src/bot/kucoin_interface.py
@@ -31,7 +31,7 @@ class KucoinInterface:
         self.user_client = User(key=API_KEY, secret=API_SECRET, passphrase=API_PASSPHRASE, url=URL)
         self.market_client = Market(key=API_KEY, secret=API_SECRET, passphrase=API_PASSPHRASE, url=URL)
 
-    def execute_trade(self, latest_bitcoin_price: float, bitcoin_holding_percentage_request: float) -> None:
+    def execute_trade(self, latest_bitcoin_price: float, bitcoin_holding_percentage_request: int) -> None:
         portfolio_breakdown = self.get_portfolio_breakdown()
 
         if bitcoin_holding_percentage_request < 0 or bitcoin_holding_percentage_request > 100:
